+++ R/data_arrange.R
@@ -1,0 +1,69 @@
+#' Arrange rows by column values
+#'
+#' `data_arrange()` orders the rows of a data frame by the values of selected
+#' columns.
+#'
+#' @param data A data frame, or an object that can be coerced to a data frame.
+#' @param ... Quoted column names. Use a dash just before column name to arrange
+#'   in decreasing order, for example "-x1".
+#' @param safe Do not throw an error if one of the variables specified doesn't
+#'   exist.
+#'
+#' @return A data frame.
+#'
+#' @examples
+#' # Arrange using several variables
+#' data_arrange(head(mtcars), "gear", "carb")
+#'
+#' # Arrange in decreasing order
+#' data_arrange(head(mtcars), "-carb")
+#'
+#' # Throw an error if one of the variables specified doesn't exist
+#' data_arrange(head(mtcars), "gear", "foo", safe = FALSE)
+#'
+#' @export
+
+data_arrange <- function(data, ..., safe = TRUE) {
+
+  el <- c(...)
+  if (length(el) == 0) return(data)
+
+  # find which vars should be decreasing
+  desc <- el[startsWith(el, "-")]
+  desc <- gsub("^-", "", desc)
+  el <- gsub("^-", "", el)
+
+  # check for variables that are not in data
+  dont_exist <- el[which(!el %in% names(data))]
+  if (length(dont_exist) > 0) {
+    if (!safe) {
+      stop(insight::format_message(
+        paste0(
+          "The following column(s) don't exist in the dataset: ",
+          text_concatenate(dont_exist), "."
+        )
+      ), call. = FALSE)
+    }
+    el <- el[-which(el %in% dont_exist)]
+  }
+
+  if (length(el) == 0) return(data)
+
+  out <- data
+
+  # reverse order for variables that should be decreasing
+  if (length(desc) > 0) {
+    for (i in desc) {
+      out[[i]] <- -xtfrm(out[[i]])
+    }
+  }
+
+  # apply ordering
+  if (length(el) == 1) {
+    data[order(out[[el]]), ]
+  } else {
+    data[do.call(order, out[, el]), ]
+  }
+}
+
+